+++ ta_lib/indicators/trend/src/ema.rs
@@ -1,6 +1,6 @@
 use core::series::Series;
 
-pub fn ema(source: &[f64], period: usize) -> Series<f64> {
+pub fn ema(source: &[f32], period: usize) -> Series<f32> {
     let source = Series::from(source);
 
     let ema = source.ema(period);
@@ -17,7 +17,7 @@ mod tests {
         let source = vec![1.0, 2.0, 3.0, 4.0, 5.0];
         let expected = vec![1.0, 1.5, 2.25, 3.125, 4.0625];
 
-        let result: Vec<f64> = ema(&source, 3).into();
+        let result: Vec<f32> = ema(&source, 3).into();
 
         assert_eq!(result, expected);
     }
