+++ ta_lib/indicators/trend/src/wma.rs
@@ -1,6 +1,6 @@
 use core::series::Series;
 
-pub fn wma(source: &[f64], period: usize) -> Series<f64> {
+pub fn wma(source: &[f32], period: usize) -> Series<f32> {
     let source = Series::from(&source);
 
     let wma = source.wma(period);
@@ -19,7 +19,7 @@ mod tests {
         let epsilon = 0.001;
         let expected = vec![0.0, 0.0, 2.333333, 3.333333, 4.333333];
 
-        let result: Vec<f64> = wma(&source, period).into();
+        let result: Vec<f32> = wma(&source, period).into();
 
         for i in 0..source.len() {
             assert!(
