+++ ta_lib/patterns/candlestick/src/bottle.rs
@@ -1,6 +1,6 @@
 use core::series::Series;
 
-pub fn bullish(open: &[f64], low: &[f64], close: &[f64]) -> Series<bool> {
+pub fn bullish(open: &[f32], low: &[f32], close: &[f32]) -> Series<bool> {
     let open = Series::from(open);
     let low = Series::from(low);
     let close = Series::from(close);
@@ -12,7 +12,7 @@ pub fn bullish(open: &[f64], low: &[f64], close: &[f64]) -> Series<bool> {
         & close.shift(1).gt(&close.shift(2))
 }
 
-pub fn bearish(open: &[f64], high: &[f64], close: &[f64]) -> Series<bool> {
+pub fn bearish(open: &[f32], high: &[f32], close: &[f32]) -> Series<bool> {
     let open = Series::from(open);
     let high = Series::from(high);
     let close = Series::from(close);
