+++ ta_lib/patterns/candlestick/src/doji.rs
@@ -1,13 +1,13 @@
 use core::series::Series;
 
-pub fn bullish(open: &[f64], close: &[f64]) -> Series<bool> {
+pub fn bullish(open: &[f32], close: &[f32]) -> Series<bool> {
     let open = Series::from(open);
     let close = Series::from(close);
 
     close.gt(&open) & close.shift(1).eq(&open.shift(1)) & close.shift(2).lt(&open.shift(2))
 }
 
-pub fn bearish(open: &[f64], close: &[f64]) -> Series<bool> {
+pub fn bearish(open: &[f32], close: &[f32]) -> Series<bool> {
     let open = Series::from(open);
     let close = Series::from(close);
