+++ ta_lib/patterns/candlestick/src/doji_double.rs
@@ -1,6 +1,6 @@
 use core::series::Series;
 
-pub fn bullish(open: &[f64], close: &[f64]) -> Series<bool> {
+pub fn bullish(open: &[f32], close: &[f32]) -> Series<bool> {
     let open = Series::from(open);
     let close = Series::from(close);
 
@@ -9,7 +9,7 @@ pub fn bullish(open: &[f64], close: &[f64]) -> Series<bool> {
         & close.shift(3).lt(&open.shift(3))
 }
 
-pub fn bearish(open: &[f64], close: &[f64]) -> Series<bool> {
+pub fn bearish(open: &[f32], close: &[f32]) -> Series<bool> {
     let open = Series::from(open);
     let close = Series::from(close);
