+++ ta_lib/patterns/candlestick/src/hammer.rs
@@ -1,6 +1,6 @@
 use core::series::Series;
 
-pub fn bullish(open: &[f64], high: &[f64], close: &[f64]) -> Series<bool> {
+pub fn bullish(open: &[f32], high: &[f32], close: &[f32]) -> Series<bool> {
     let open = Series::from(open);
     let high = Series::from(high);
     let close = Series::from(close);
@@ -15,7 +15,7 @@ pub fn bullish(open: &[f64], high: &[f64], close: &[f64]) -> Series<bool> {
         & body.shift(2).lt(&body.shift(3))
 }
 
-pub fn bearish(open: &[f64], low: &[f64], close: &[f64]) -> Series<bool> {
+pub fn bearish(open: &[f32], low: &[f32], close: &[f32]) -> Series<bool> {
     let open = Series::from(open);
     let low = Series::from(low);
     let close = Series::from(close);
